SpecificDayViewModel.kt
